+++ test/doubler_test_objects/double_me.gd
@@ -11,3 +11,6 @@ func has_one_param(one):
 
 func has_two_params_one_default(one, two=null):
     pass
+
+func get_position():
+    return .get_position()